requirements.txt
opencv-python-headless==4.8.1.78
streamlit_webrtc==0.47.1
twilio~=8.10.0
Pillow
Numpy
mediapipe